支持加载用例阶段指定用例所在模块

+++ python_unittest/tests/conftest.py
@@ -1,20 +1,26 @@
 import os
 import glob
+from pathlib import Path
 import pytest
 import shutil
-from .test_run import testdata_dir
+
+testdata_dir: str = str(Path(__file__).parent.absolute().joinpath("testdata"))
+
 
 @pytest.hookimpl(tryfirst=True, hookwrapper=True)
 def pytest_sessionfinish(session, exitstatus):
-    json_files = glob.glob(os.path.join(testdata_dir, '*.json'))
-    
+    json_files = glob.glob(os.path.join(testdata_dir, "*.json"))
+
     for json_file in json_files:
         try:
             os.remove(json_file)
         except OSError as e:
             print(f"Error removing file {json_file} err: {e}")
     # 删除结果文件
-    os.remove("test_results.xml")
+    if os.path.exists("test_results.xml"):
+        os.remove("test_results.xml")
     # 删除coverage文件
-    shutil.rmtree(os.path.join(testdata_dir, "testsolar_coverage"))
+    coverage_path = os.path.join(testdata_dir, "testsolar_coverage")
+    if os.path.exists(coverage_path):
+        shutil.rmtree(coverage_path)
     yield
